fix: handle json-rpc error object
